update exmaples/basic module

+++ examples/basic/src/db/user.rs
@@ -4,7 +4,7 @@ use lan_be_frame::db::Entity;
 #[table_name = "user"]
 pub struct User {
     #[primary_key]
-    pub id: u16,
+    pub id: i32,
     pub name: String,
     pub email: String,
 }
